+++ socket/src/main/java/io/eagle/config/KafkaConsumerConfig.java
@@ -2,8 +2,7 @@ package io.eagle.config;
 
 import com.google.common.collect.ImmutableMap;
 import io.eagle.common.KafkaConstants;
-import io.eagle.domain.order.dto.BroadcastMessageDto;
-import io.eagle.domain.order.dto.MessageDto;
+import io.eagle.domain.order.dto.response.BroadcastMessageDto;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.apache.kafka.common.serialization.StringDeserializer;
 import org.apache.kafka.common.serialization.StringSerializer;
