+++ socket/src/main/java/io/eagle/domain/order/controller/OrderController.java
@@ -1,9 +1,9 @@
 package io.eagle.domain.order.controller;
 
 import io.eagle.common.KafkaConstants;
-import io.eagle.domain.order.dto.ErrorDto;
-import io.eagle.domain.order.dto.MessageDto;
-import io.eagle.domain.order.dto.BroadcastMessageDto;
+import io.eagle.domain.order.dto.response.ErrorDto;
+import io.eagle.domain.order.dto.request.MessageDto;
+import io.eagle.domain.order.dto.response.BroadcastMessageDto;
 import io.eagle.domain.order.service.OrderService;
 import io.eagle.exception.SocketException;
 import lombok.RequiredArgsConstructor;
@@ -23,6 +23,7 @@ public class OrderController {
 
     @MessageMapping("/purchase") //   url : "order/purchase" 로 들어오는 정보 처리
     public void purchase(MessageDto message){
+        message.setRequesterId(1L); // TODO : 임시 사용자
         BroadcastMessageDto broadcastMessageDto = orderService.purchaseMarket(message);
         log.info("[STOMP Producer] user purchase market {} price : {}, amount : {}, left : {}",message.getMarketId(),message.getPrice(), message.getAmount(), broadcastMessageDto.getAmount());
         kafkaTemplate.send(KafkaConstants.KAFKA_TOPIC, broadcastMessageDto);
