+++ socket/src/main/java/io/eagle/domain/order/dto/response/ErrorDto.java
@@ -1,4 +1,4 @@
-package io.eagle.domain.order.dto;
+package io.eagle.domain.order.dto.response;
 
 import lombok.Builder;
 import lombok.Data;
